Overhauled Warp system, added player health and added to A*
The warp system is no longer as convoluted as it used to be. A player touches the trigger and it disables all over warps and activates it. If the player's health reaches 0 then they go back to the warp. A* algorithm has been modified to be more extensible to the units it controls.